Created a utils folder and moved all js utils
